fix(db): align classifier privilege and tie validation

+++ supabase/migrations/20260823152435_optimize_storefront_cluster_guide_classifier_rpc.sql
@@ -98,7 +98,7 @@ BEGIN
       OR EXISTS (
         SELECT 1
         FROM pg_catalog.jsonb_array_elements(p_cluster_rules) AS rule(value)
-        GROUP BY rule.value ->> 'rule_order'
+        GROUP BY (rule.value ->> 'rule_order')::integer
         HAVING pg_catalog.count(*) > 1
       )
       OR EXISTS (
